Productos

+++ Frontend_Inventario/Modelos/Producto_Modelo.cs
@@ -7,7 +7,7 @@
         public string descripcion { get; set; } = string.Empty;
         public decimal precioCompra { get; set; }
         public decimal precioVenta { get; set; }
-        public int margenGanancia { get; set; }
+        public decimal margenGanancia { get; set; }
         public int stockActual { get; set; }
         public int stockMinimo { get; set; }
         public int idCategoria { get; set; }
@@ -23,7 +23,7 @@
         public string descripcion { get; set; } = string.Empty;
         public decimal precioCompra { get; set; }
         public decimal precioVenta { get; set; }
-        public int margenGanancia { get; set; }
+        public decimal margenGanancia { get; set; }
         public int stockActual { get; set; }
         public int stockMinimo { get; set; }
         public int idCategoria { get; set; }
